hack_book_agent/hack_library_agent/requirements.txt
a2a-sdk
google-adk
python-dotenv
toolbox-core
uvicorn
